Fix and simplify filter_adjlist.
The docs say it should remove elements in insertion order, but it
doesn't because `set` is unordered. It is dependent on the order Python
decides, which does in fact change between versions, and the related
test breaks on 3.8.

The test itself is also wrong because removed elements in a 'nice'
manner, not insertion order.